src/requirements.txt
cffi==1.7.0
cryptography==1.4
defusedxml==0.4.1
Django==1.9.4
django-allauth==0.25.2
django-appconf==1.0.2
django-autofixture==0.12.0
django-extensions==1.6.1
django-filter==0.12.0
django-imagekit==3.3
django-reversion==2.0.6
django-webpack-loader==0.3.3
djangorestframework==3.4.3
Fabric==1.10.2
flake8==2.5.4
idna==2.1
mccabe==0.4.0
numpy==1.11.0
oauthlib==1.1.2
paramiko==2.0.2
pep8==1.7.0
pilkit==1.1.13
Pillow==3.3.0
psycopg2==2.6.1
pyasn1==0.1.9
pycparser==2.14
pyflakes==1.0.0
python3-openid==3.0.10
requests==2.10.0
requests-oauthlib==0.6.2
six==1.10.0
